close (#15)
* feat: add resume command

* test: add command processing tests

* fix: move script.js to js/ directory

* fix: move script.js to js/ directory

* fix: ug move script clean stuff

---------
