+++ .gitignore
@@ -6,3 +6,5 @@ _site/
 .bundle/
 vendor/
 terminal-portfolio
+mdriscoll93.github.io.code-workspace
+node_modules/
