+++ __tests__/script.test.js
@@ -1,0 +1,14 @@
+const { processCommand } = require('../script');
+
+describe('processCommand', () => {
+  test('help command outputs list', () => {
+    expect(processCommand('help')).toBe('Available commands: help, cat, resume, clear');
+  });
+
+  test('clear command clears terminal', () => {
+    const terminal = { clear: jest.fn() };
+    const result = processCommand('clear', { terminal });
+    expect(terminal.clear).toHaveBeenCalled();
+    expect(result).toBe('');
+  });
+});
