+++ package.json
@@ -5,5 +5,11 @@
     "node-pty": "^1.0.0",
     "socket.io": "^4.8.1",
     "xterm": "^5.3.0"
+  },
+  "devDependencies": {
+    "jest": "^29.6.1"
+  },
+  "scripts": {
+    "test": "jest"
   }
 }
